Create create_archive function + associated types
Adds a new module called `api`. This module is designed to mimick the
high level of functionality of `bita` but make to make it available
programmatically using `bitar`.

Currently, only one method is implemented, which is `create_archive`.
This is the same logic used from the bita `bita compress` command with
some logic cleaned up and some changes to make it suitable as a library
function.

This change moves `tempfile` from dev dependency to a hard dependency in
`bitar`. This is needed to write archive data to a temporary file before
prepending the header.