Bug: Fix local paths incorrectly parsed as urls on windows in bita
Previously in bita, when a user provided an input path an attempt was
made to parse it as a URL. If this failed it would fallback to being
used as a local path. However, windows paths are succesfully parsed as a
url (e.g. `C:\temp\my_archive.cba`) which caused a bug on windows where
inputs were always treated as a remote path.

This fix inverts the behavior and first checks if the path is a valid
local path before checking if it's a valid remote path.